move convert schema to utils
