+++ server/src/collections/collections.service.ts
@@ -21,12 +21,16 @@ import {
   CompactReq,
   HasCollectionReq,
   CountReq,
-  convertToDataType,
   FieldSchema,
-  FieldType,
 } from '@zilliz/milvus2-sdk-node';
 import { Parser } from '@json2csv/plainjs';
-import { throwErrorFromSDK, findKeyValue, genRows, ROW_COUNT } from '../utils';
+import {
+  throwErrorFromSDK,
+  findKeyValue,
+  genRows,
+  ROW_COUNT,
+  convertFieldSchemaToFieldType,
+} from '../utils';
 import { QueryDto, ImportSampleDto, GetReplicasDto } from './dto';
 import { CollectionData } from '../types';
 import { SchemaService } from '../schema/schema.service';
@@ -374,33 +378,14 @@ export class CollectionsService {
       enable_dynamic_field: !!collection.enable_dynamic_field,
     };
 
-    if (collection.schema.fields.some((f: FieldSchema) => f.is_partition_key)) {
+    if (
+      collection.schema.fields.some((f: FieldSchema) => f.is_partition_key) &&
+      collection.num_partitions
+    ) {
       createCollectionParams.num_partitions = Number(collection.num_partitions);
     }
 
     console.dir(createCollectionParams, { depth: null });
     return await this.createCollection(createCollectionParams);
   }
-}
-
-function convertFieldSchemaToFieldType(fieldSchema: FieldSchema) {
-  const fieldType: FieldType = {
-    name: fieldSchema.name,
-    description: fieldSchema.description,
-    data_type: convertToDataType(fieldSchema.data_type),
-    element_type: convertToDataType(fieldSchema.element_type),
-    is_primary_key: fieldSchema.is_primary_key,
-    is_partition_key: fieldSchema.is_partition_key,
-    autoID: fieldSchema.autoID,
-  };
-
-  // Convert type_params from array to object
-  if (fieldSchema.type_params) {
-    fieldType.type_params = {};
-    for (const param of fieldSchema.type_params) {
-      fieldType.type_params[param.key] = param.value;
-    }
-  }
-
-  return fieldType;
 }
